updated frontend

+++ src/Pages/Details/MovieDetails.jsx
@@ -3,10 +3,11 @@ import { Helmet } from "react-helmet-async";
 import { FaHeartCirclePlus } from "react-icons/fa6";
 import { RiDeleteBin6Fill } from "react-icons/ri";
 import { RxUpdate } from "react-icons/rx";
-
 import { Link, useLoaderData, useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import { AuthContext } from "../../Auth/AuthProvider";
+import Swal from "sweetalert2";
+
 const MovieDetails = () => {
   const movieData = useLoaderData();
   const { user } = useContext(AuthContext);
@@ -15,16 +16,32 @@ const MovieDetails = () => {
   const navigate = useNavigate();
 
   const handleDeleteBtn = (id) => {
-    fetch(`https://movie-protal-server.vercel.app/all-movies/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          toast.success("Movie successfully deleted");
-          navigate("/all-movies");
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`https://movie-protal-server.vercel.app/all-movies/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+              navigate("/all-movies");
+            }
+          });
+      }
+    });
   };
   const handleFavBtn = () => {
     const favMovie = {
@@ -54,7 +71,7 @@ const MovieDetails = () => {
       <Helmet>
         <title>Details-Movie Portal</title>
       </Helmet>
-      <div className="container w-10/12 mx-auto mb-20">
+      <div className="container w-10/12 mx-auto mb-20 mt-28">
         <div className="p-10 bg-base-200 rounded-2xl">
           <div className="flex flex-col lg:flex-row gap-10 items-center">
             <img src={poster} className="w-[50%] rounded-lg shadow-2xl" />
@@ -97,19 +114,19 @@ const MovieDetails = () => {
               <div className="flex flex-col md:flex-row gap-5 flex-wrap pt-10 ">
                 <button
                   onClick={() => handleDeleteBtn(_id)}
-                  className="btn btn-primary  dark:text-white"
+                  className="btn bg-light-primary hover:bg-dark-primary text-white"
                 >
                   Delete <RiDeleteBin6Fill className="text-lg" />
                 </button>
                 <button
                   onClick={handleFavBtn}
-                  className="btn btn-primary  dark:text-white"
+                  className="btn bg-light-primary hover:bg-dark-primary  text-white"
                 >
                   Add To Favorite <FaHeartCirclePlus className="text-lg" />
                 </button>
                 <Link
                   to={`/update/${_id}`}
-                  className="btn btn-primary  dark:text-white"
+                  className="btn bg-light-primary hover:bg-dark-primary text-white"
                 >
                   Update Movie <RxUpdate className="text-lg" />
                 </Link>
